20139 - organize aws step_function
