+++ airflow/providers/amazon/aws/sensors/step_function_execution.py
@@ -15,74 +15,14 @@
 # specific language governing permissions and limitations
 # under the License.
 
-import json
-from typing import Optional
+"""This module is deprecated. Please use :mod:`airflow.providers.amazon.aws.sensors.step_function`."""
 
-from airflow.exceptions import AirflowException
-from airflow.providers.amazon.aws.hooks.step_function import StepFunctionHook
-from airflow.sensors.base import BaseSensorOperator
+import warnings
 
+from airflow.providers.amazon.aws.sensors.step_function import StepFunctionExecutionSensor  # noqa
 
-class StepFunctionExecutionSensor(BaseSensorOperator):
-    """
-    Asks for the state of the Step Function State Machine Execution until it
-    reaches a failure state or success state.
-    If it fails, failing the task.
-
-    On successful completion of the Execution the Sensor will do an XCom Push
-    of the State Machine's output to `output`
-
-    :param execution_arn: execution_arn to check the state of
-    :type execution_arn: str
-    :param aws_conn_id: aws connection to use, defaults to 'aws_default'
-    :type aws_conn_id: str
-    """
-
-    INTERMEDIATE_STATES = ('RUNNING',)
-    FAILURE_STATES = (
-        'FAILED',
-        'TIMED_OUT',
-        'ABORTED',
-    )
-    SUCCESS_STATES = ('SUCCEEDED',)
-
-    template_fields = ['execution_arn']
-    template_ext = ()
-    ui_color = '#66c3ff'
-
-    def __init__(
-        self,
-        *,
-        execution_arn: str,
-        aws_conn_id: str = 'aws_default',
-        region_name: Optional[str] = None,
-        **kwargs,
-    ):
-        super().__init__(**kwargs)
-        self.execution_arn = execution_arn
-        self.aws_conn_id = aws_conn_id
-        self.region_name = region_name
-        self.hook: Optional[StepFunctionHook] = None
-
-    def poke(self, context):
-        execution_status = self.get_hook().describe_execution(self.execution_arn)
-        state = execution_status['status']
-        output = json.loads(execution_status['output']) if 'output' in execution_status else None
-
-        if state in self.FAILURE_STATES:
-            raise AirflowException(f'Step Function sensor failed. State Machine Output: {output}')
-
-        if state in self.INTERMEDIATE_STATES:
-            return False
-
-        self.log.info('Doing xcom_push of output')
-        self.xcom_push(context, 'output', output)
-        return True
-
-    def get_hook(self) -> StepFunctionHook:
-        """Create and return a StepFunctionHook"""
-        if self.hook:
-            return self.hook
-
-        self.hook = StepFunctionHook(aws_conn_id=self.aws_conn_id, region_name=self.region_name)
-        return self.hook
+warnings.warn(
+    "This module is deprecated. Please use `airflow.providers.amazon.aws.sensors.step_function`.",
+    DeprecationWarning,
+    stacklevel=2,
+)
